% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-class.R
\name{task_update_call_result}
\alias{task_update_call_result}
\title{Update a task with any finished call result from a persistent session.}
\usage{
task_update_call_result(task)
}
\description{
Purpose:
\itemize{
\item Detect when a \code{callr::r_session} call has completed even though the
underlying R process remains alive and reusable.
}
}
\details{
Parameters:
\itemize{
\item \code{task}: Internal \code{Task} object.
}

Returns:
\itemize{
\item Invisibly returns the task object.
}
}
\keyword{internal}
